+++ tests/test_ophys/test_tools_roiextractors.py
@@ -4,6 +4,7 @@ from datetime import datetime
 from pathlib import Path
 from tempfile import mkdtemp
 from types import MethodType
+import math
 from typing import List, Literal, Optional, Tuple
 from unittest.mock import Mock
 
@@ -1469,7 +1470,7 @@ class TestAddPhotonSeries(TestCase):
         available_memory_in_bytes = psutil.virtual_memory().available
 
         excess = 1.5  # Of what is available in memory
-        num_frames_to_overflow = (available_memory_in_bytes * excess) / (element_size_in_bytes * np.prod(image_size))
+        num_frames_to_overflow = (available_memory_in_bytes * excess) / (element_size_in_bytes * math.prod(image_size))
 
         # Mock recording extractor with as much frames as necessary to overflow memory
         mock_imaging = Mock()